Refactor existing block into modular directory structure
Move player ratings block into src/player-ratings/ folder to prepare
for future multi-block architecture. Extract Edit component into
separate file for better maintainability. Build paths remain
unchanged to preserve compatibility.
