+++ src/frontend.js
@@ -1,18 +1,2 @@
-/* global navigator */
-// Frontend JavaScript for Pickleball Ratings Block
-window.pbrCopyToClipboard = function ( duprId, button ) {
-	const icon = button.querySelector( '.dashicons' );
-
-	// Copy to clipboard
-	navigator.clipboard.writeText( duprId ).then( function () {
-		// Change icon and tooltip
-		icon.className = 'dashicons dashicons-yes';
-		button.title = 'Copied!';
-
-		// Reset after 2 seconds
-		setTimeout( function () {
-			icon.className = 'dashicons dashicons-clipboard';
-			button.title = 'Copy DUPR ID: ' + duprId;
-		}, 2000 );
-	} );
-};
+// Import frontend scripts for all blocks
+import './player-ratings/frontend';
